refactor(civic-center-service): PR comment resolution
