+++ backend/src/modules/_publicAPI/public-api.module.ts
@@ -3,7 +3,8 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { ApplicationModule } from '../application/application.module';
 import { CivicCenterModule } from '../civic-center-service/civic-center.module';
 import { OrganizationModule } from '../organization/organization.module';
-import { PublicAPIController } from './public-api.controller';
+import { CivicCenterPublicController } from './controllers/civic-center-public.controller';
+import { PublicAPIController } from './controllers/public-api.controller';
 import { PublicKeysManager } from './public-keys-manager.service';
 import { PublicKeys } from './public-keys.entity';
 import { CivicCenterPublicService } from './services/civic-center-public.service';
@@ -15,7 +16,7 @@ import { CivicCenterPublicService } from './services/civic-center-public.service
     ApplicationModule,
     CivicCenterModule,
   ],
-  controllers: [PublicAPIController],
+  controllers: [PublicAPIController, CivicCenterPublicController],
   providers: [PublicKeysManager, CivicCenterPublicService],
   exports: [],
 })
